FLAS-40-Cards refactor: -change folder from pages to features
